Dice values
Assigning and drawing dice values

+++ Assets/Scripts/Dice/Data/DiceValue.cs
@@ -5,17 +5,20 @@ namespace Dice.Data
     [CreateAssetMenu(fileName = "DiceValue", menuName = "ScriptableObjects/DiceValue", order = 0)]
     public class DiceValue : ScriptableObject
     {
-        [Tooltip("This name will be displayed to user after getting a dice throw result.")]
-        [SerializeField] private string displayName;
-        
-        [Header("Presentation - this value will appear on dice")]
-        [SerializeField] private string text;
-        [SerializeField] private float fontSize;
+        [Tooltip("This name will be displayed to user after getting a dice throw result.")] 
+        [SerializeField]
+        private string displayName;
+
+        [Header("Presentation - this value will appear on dice")] 
+        [SerializeField]
+        private string text;
+
+        [SerializeField] private int fontSize;
 
         public string DisplayName => displayName;
 
         public string Text => text;
 
-        public float FontSize => fontSize;
+        public int FontSize => fontSize;
     }
 }
